Restructure frame codebase
