+++ agtern/gui/frames/__init__.py
@@ -1,8 +1,3 @@
-from .internship_detail_frame import InternshipDetailFrame
-from .internship_entry_frame import InternshipEntryFrame
-from .internship_list_frame import InternshipListFrame
-from .internship_search_frame import InternshipSearchFrame
-from .settings_frame import SettingsFrame
-from .top_bar_frame import TopBarFrame
-from .vertical_scrolled_frame import VerticalScrolledFrame
-from .home_frame import HomeFrame
+from .home import *
+from .settings import *
+from .top_bar import *
