+++ agtern/gui/frames/home/detail/__init__.py
@@ -1,0 +1,2 @@
+from .card import *
+from .frame import InternshipDetailFrame
