+++ agtern/gui/frames/home/detail/card/apply_button/__init__.py
@@ -1,0 +1,1 @@
+from .frame import InternshipApplyButton
