+++ agtern/gui/frames/home/list/__init__.py
@@ -1,0 +1,2 @@
+from .frame import InternshipListFrame
+from .scrolled import *
