+++ agtern/gui/frames/home/list/scrolled/entry/__init__.py
@@ -1,0 +1,1 @@
+from .frame import InternshipEntryFrame
