+++ agtern/gui/frames/home/search/__init__.py
@@ -1,0 +1,1 @@
+from .frame import InternshipSearchFrame
