+++ agtern/gui/frames/home/search/frame.py
@@ -2,7 +2,7 @@ import tkinter as tk
 import tkinter.ttk as ttk
 from textwrap import fill
 
-from .settings_frame import SettingsFrame
+from ...settings import SettingsFrame
 
 
 # A dict, keyed by association
