+++ agtern/gui/frames/settings/__init__.py
@@ -1,0 +1,1 @@
+from .frame import SettingsFrame
