+++ agtern/gui/frames/settings/frame.py
@@ -2,7 +2,7 @@ import tkinter as tk
 from datetime import datetime
 from tkinter import ttk
 
-from ...common import LOG
+from ....common import LOG
 
 
 class SettingsFrame(tk.Frame):
