+++ agtern/gui/frames/top_bar/__init__.py
@@ -1,0 +1,1 @@
+from .frame import TopBarFrame
